requirements.txt
fastapi 
uvicorn 
python-multipart
pydantic[email]
python-dotenv